fix: issue-66
